+++ tests/test_datasources.py
@@ -10,6 +10,98 @@ from laktory._testing import spark
 
 paths = Paths(__file__)
 
+schema = {
+    "type": "struct",
+    "fields": [
+        {
+            "name": "data",
+            "type": {
+                "type": "struct",
+                "fields": [
+                    {"name": "@id", "type": "string", "nullable": True, "metadata": {}},
+                    {
+                        "name": "_created_at",
+                        "type": "string",
+                        "nullable": True,
+                        "metadata": {},
+                    },
+                    {
+                        "name": "_name",
+                        "type": "string",
+                        "nullable": True,
+                        "metadata": {},
+                    },
+                    {
+                        "name": "_producer_name",
+                        "type": "string",
+                        "nullable": True,
+                        "metadata": {},
+                    },
+                    {
+                        "name": "close",
+                        "type": "double",
+                        "nullable": True,
+                        "metadata": {},
+                    },
+                    {
+                        "name": "created_at",
+                        "type": "string",
+                        "nullable": True,
+                        "metadata": {},
+                    },
+                    {
+                        "name": "high",
+                        "type": "double",
+                        "nullable": True,
+                        "metadata": {},
+                    },
+                    {"name": "low", "type": "double", "nullable": True, "metadata": {}},
+                    {
+                        "name": "open",
+                        "type": "double",
+                        "nullable": True,
+                        "metadata": {},
+                    },
+                    {
+                        "name": "symbol",
+                        "type": "string",
+                        "nullable": True,
+                        "metadata": {},
+                    },
+                ],
+            },
+            "nullable": True,
+            "metadata": {},
+        },
+        {"name": "description", "type": "string", "nullable": True, "metadata": {}},
+        {"name": "name", "type": "string", "nullable": True, "metadata": {}},
+        {
+            "name": "producer",
+            "type": {
+                "type": "struct",
+                "fields": [
+                    {
+                        "name": "description",
+                        "type": "string",
+                        "nullable": True,
+                        "metadata": {},
+                    },
+                    {
+                        "name": "name",
+                        "type": "string",
+                        "nullable": True,
+                        "metadata": {},
+                    },
+                    {"name": "party", "type": "long", "nullable": True, "metadata": {}},
+                ],
+            },
+            "nullable": True,
+            "metadata": {},
+        },
+    ],
+}
+
+
 # DataFrame
 pdf = pd.DataFrame(
     {
@@ -38,6 +130,73 @@ def test_file_data_source_read():
     source = FileDataSource(
         path=os.path.join(paths.data, "./events/yahoo-finance/stock_price"),
         as_stream=False,
+    )
+    df = source.read(spark)
+    assert df.count() == 80
+    assert df.columns == [
+        "data",
+        "description",
+        "name",
+        "producer",
+    ]
+
+    return df
+
+
+def test_file_data_source_read_jsonl():
+    source = FileDataSource(
+        path=os.path.join(paths.data, "./events/yahoo-finance/stock_price"),
+        format="JSONL",
+    )
+    df = source.read(spark)
+    assert df.count() == 80
+    assert df.columns == [
+        "data",
+        "description",
+        "name",
+        "producer",
+    ]
+
+    return df
+
+
+def test_file_data_source_read_schema():
+
+    # Schema as dict
+    source = FileDataSource(
+        path=os.path.join(paths.data, "./events/yahoo-finance/stock_price"),
+        as_stream=False,
+        schema=schema,
+    )
+    df = source.read(spark)
+    assert df.count() == 80
+    assert df.columns == [
+        "data",
+        "description",
+        "name",
+        "producer",
+    ]
+
+    # Schema as list
+    source = FileDataSource(
+        path=os.path.join(paths.data, "./events/yahoo-finance/stock_price"),
+        as_stream=False,
+        schema=[f for f in schema["fields"]],
+    )
+    df = source.read(spark)
+    assert df.count() == 80
+    assert df.columns == [
+        "data",
+        "description",
+        "name",
+        "producer",
+    ]
+
+    # Schema as list
+    source = FileDataSource(
+        path=os.path.join(paths.data, "./events/yahoo-finance/stock_price"),
+        as_stream=False,
+        schema="data STRUCT<_created_at STRING, _name STRING, _producer_name STRING, close DOUBLE, created_at STRING, high DOUBLE, low DOUBLE, open DOUBLE, symbol STRING>, description STRING, name STRING, producer STRUCT<description STRING, name STRING, party LONG>",
     )
     df = source.read(spark)
     assert df.count() == 80
@@ -135,6 +294,8 @@ def test_table_data_source():
 if __name__ == "__main__":
     test_file_data_source()
     test_file_data_source_read()
+    test_file_data_source_read_jsonl()
+    test_file_data_source_read_schema()
     test_file_data_source_polars()
     test_memory_data_source()
     test_table_data_source()